Remove @RequiresAPI annotations.
This annotation requires adding a new dependency to Tink. Normally this should be fine, but in in this case it has many ripple effects. For example, it requires Android users to change their Gradle config, i.e., upgrading compileSdkVersion to 28 and adding android.useAndroidX=true.

We understand that AndroidX is the future of Android libraries, and it'd inevitably be used in Tink. However, this annotation is not too important for us to start depending on AndroidX for now.

PiperOrigin-RevId: 315809861