Discontinue Java 7 support.
The primary reason to drop that support is because Bazel [1] and Protobuf [2] have stopped supporting Java 7.

The second reason is that nobody in the wild should rely on Java 7 any more because it was discontinued years ago.

The third reason is, due to the first reason, Tink 1.4.0-rc1 and 1.4.0-rc2 were accidentally compiled with Java 8. No user, however, has complained.

Remove the JAVACOPTS option. Because Tink depends on Protobuf, Tink cannot really specify which Java version it can support, but it has to use whatever version that Protobuf supports.

[1] https://github.com/bazelbuild/bazel/pull/9450
[2] https://github.com/protocolbuffers/protobuf/pull/7306

PiperOrigin-RevId: 321114818