Create Bazel setup for TypeScript/JavaScript
Open source users may now fully build Tink and run the tests, and we do so (testing against Chrome and Firefox) in CI. This also unblocks publishing it to npm.

ng_package.bzl is a rule that creates an npm package. It's bitrotted a little, so we fork from the one in the Angular repo in order to make it build. It's under a top-level third_party directory in order to comply with https://opensource.google/docs/releasing/preparing/#third-party-components.

PiperOrigin-RevId: 328226655