Change the primitive wrapper such that it can wrap a primitive of type A into a primitive of type B.
This is, strictly speaking, incompatible with previously releases. However, the incompatibility will only manifest itself if a user uses their own primitive, which I don't expect to be the case.

Furthermore, the fix is very easy (see e.g. the change in AeadWrapper). Hence, in the spirit of not having much technical debt, I suggest to be backwards incompatible.

PiperOrigin-RevId: 331095397