This change removes TypeScript code that cannot be reached.
Unreachable code is error prone: it can suggest that the author of the code did not quite understand their code's control flow, or control flow changed in refactorings, etc. At best, the code is misleading.

TypeScript compiler includes an option to strictly check unreachable code (allowUnreachableCode: false). We want to enable this option in google3, but must first clean up any code that violates it.

This code was only statically detected to be unreachable. If your code has incorrect type annotations, there's a very small chance the code is still executed.

PiperOrigin-RevId: 334189859