Retry one more time when cannot encrypt or decrypt with a KeyStore key.
Android KeyStore has a hard limitation of 15 key operations that can be in flight at any given time (4 for Strongbox). If this limit is exceeded the least recently key usage operation gets pruned. This does not affect key generation or attestation though.

When encrypting or decrypting with a KeyStore key, if Tink encountered an error, it'd retry one more time. We can't retry more than once because we can't tell why KeyStore failed -- if the error is not transient, retrying multiple times might lengthen the user's wait time.

PiperOrigin-RevId: 336095561