Fix AWS KMS test with credentials file
Because the KMS client registry just iterates over registered clients
until it finds a client where the keyURI is .Supported(), the second
test re-uses the .csv file rather than testing the .cred file.

Re-enabling the test (by explicitly clearing all KMS clients in
between iterations) reveals that the .cred file isn't formatted in
the way that the underlying Go library expects. Create an alternative
.ini version that has the expected format, leaving the existing .cred
file as-is because the Java library expects its format.

PiperOrigin-RevId: 340811610