Make VerifiedJwt not movable.
Having move constructor/assignment operator to default means that we can write something like verified_1 = std::move(verified_2), which will move the member from verified_2 to verified_1. This will change verified_2, which in turn will not really be a verified_2. After this change, above code will make a copy, and they are both the same.

PiperOrigin-RevId: 367178216