Fix code directly constructing `jspb.Message` instances
`jspb.Message` is the base class of all Apps Jspb generated message classes.  To get a valid instance it relies on a specific initialization sequence which is not satisfied by direct calls to the constructor.  This LSC will modify all code relying directly constructing this class (or otherwise relying on the constructor being instantiable) to use a concrete proto instance instead, generally the `Empty` proto.
​
When directly constructing `Message` the object is left in an inconsistent state which may trigger odd behavior when using standard functionality.  To prevent this we will be changing the constructor parameters and marking the base class `@abstract`.

PiperOrigin-RevId: 374347708