Make JWT issuer and subject validation more strict.
RFC 8725 requires these claims to be validated if they are present in the token, see https://datatracker.ietf.org/doc/html/rfc8725#section-3.8.
We already do this for audience, and I think we should be consistent and also do this for issuer and subject.

With this cl, we now reject tokens that have these claims set, but where the validator does not have them set.

These stricter rules may however be a problem in some situations: for example, it would basically be impossible to migrate from one subject to another subject or to migrate from no subject to a subject. Therefore, we also need a way to explicitly ignore the validation of these claims. This also applies to audience.

Additionally, I also change the method name from "setIssuer" to "expectIssuer". I think this is a bit more natural, since the validator does not really have an issuer, it checks the issuer.

PiperOrigin-RevId: 376328140