Fix how Cecpq2HkdfSenderKemBoringSsl stores pubkeys.
When Cecpq2HkdfSenderKemBoringSsl was written, we tried to save a couple string copies by storing only the pointers to buffers that contained X25519 and HRSS public keys instead of explicitly copying them. This could lead to undefined behavior in case the caller does not ensure the existence of the public key buffers long enough, thus we changed Cecpq2HkdfSenderKemBoringSsl to explicitly copy and store these public keys.

PiperOrigin-RevId: 376856212