Remove subject validation from JwtValidator in Java.
Best practices of JWT require the subject claim to be validated if present, see
https://datatracker.ietf.org/doc/html/rfc8725#section-3.8. Currently, we enforce
this by requiring that the verifier calls "expectSubjet(...)" for all tokens
that have a subject set. This validation can be skipped with "ignoreSubject"

But in most cases, subject can have different values. This is currently not well
supported: the verifier needs to call "ignoreSubject" and validate the subject
themselves.

So I think the current "expectSubject" function is not really helpful, and want
to remove it. The verifier needs to validate the subject themselves.

(Issuer and Audience are different, as these should normally be known by the verifier, and they should be validated when they are set in the token.)

PiperOrigin-RevId: 391538770