Use _ADDITIONAL_KEY_TEMPLATES in mac_test.
It is preferable to not pass key templates as parameters to the test cases, because
- the test output looks nicer.
- it makes it easier to replace the supported_key_types.KEY_TEMPLATE with an RPC call.

Also, I reduce the amount of test cases. I think it is enough to just test one key template for raw and one for legacy, because this is handled by the primitive wrapper which is independent of the particular key template used.

PiperOrigin-RevId: 393314432