Replace use of BN_bn2bin_padded with BN_bn2binpad for copatibility with OpenSSL.
BN_bn2binpad is defined by both OpenSSL and BoringSSL. In BoringSSL it
internally calls BN_bn2bin_padded, the difference is in the return value:
in the case of BN_bn2binpad is the length of the buffer or -1 on failure,
while BN_bn2bin_padded returns 0 on failure, and 1 on success.
Created new utility function BignumToBinaryPadded to perform common checks.

PiperOrigin-RevId: 401722883