Set the length of x and y according to the curve in ParseEcPublicKey.
When a new EcKey is generated, the pub_x and pub_y always have a constant length, so leading zeros may be present. To be consistent, it is better to also keep the leading zeros when converting a PEM key into an EcKey.

Also, add some tests:
- Add an example key that exposes the difference: the key has a leading zero byte in pub_x. Using BN_num_bytes, this leading zero would be removed.
- Add a test that generates keys exports the public key and imports them again.

NOKEYCHECK

NOKEYCHECK=True
PiperOrigin-RevId: 409071718