Update protobuf to v3.19.3.
The dependencies in examples_java_deps_init needed to be updated to work with this version of protobuf.

Also, add some explicit six requirements. This was not required until now because protobuf depends on six, and most targets imported protobuf. But the new protobuf version how does not require six anymore, but absl-py still depends on it and the build would break if I don't add this requirement explicitly. But soon, the six dependency will also be removed from absl-py.

PiperOrigin-RevId: 423124555