JWT MAC Internal Primitive.
JWT interfaces need access to the key material so that Tink can write the KID header (for TINK keys) in the JWT header. This requires two primitive interfaces. One internal that is KID aware and a public one which wraps the internal with a primitive set.

This commit implements the internal primitive.

PiperOrigin-RevId: 426279795