Require that keysets always have a primary key, even if they only contain public keys.
For hybrid encryption, this is anyways necessary, because we need the primary key to encrypt. For signatures, it would be possible to verify without a primary key, but it currently doesn't work and panics.

The panic happens in signer_factory.newWrappedVerifier, on the first line. The function tries to access ps.Primary.Primitive, but that fails because ps.Primary = nil.

I think it is best to be consistent with the C++ implementation and require a primary key.

PiperOrigin-RevId: 438507050