Allow missing KeysetInfo in JsonKeysetReader.java
Currently, the C++ KeysetHandle.write method does not set the KeysetInfo. To make sure that keysets written in C++ can be read in Java, we must allow missing KeysetInfo in JsonKeysetReader.java.

Note that the "KeysetInfo" field is anyways ignored by KeysetHandle.read, only the "EncryptedKeyset" is used to generate the KeysetHandle.

PiperOrigin-RevId: 440096586