Write unsigned key IDs in KeyInfo.
Currently, the primary key is written as unsigned int, but the key info key id is written assigned int, which is inconsistent. Also, this causes an error when the keyset is read in C++.

This fix shouldn't cause anything to fail because the KeyInfo is anyways ignored by KeysetHandle.read.

PiperOrigin-RevId: 440120826