Allow buffering and re-reading data when trying to find the correct StreamingAEAD key, but don't keep the buffer around once the key is found.
This replaces the previous implementation that buffered the entire stream into
a `bytes.Buffer` using `io.TeeReader`. The buffer could be re-read when trying
a new key, but it was still referenced and written to when a key was found and
it was no longer needed. That caused the entire stream to be held in the buffer
instead of just the current block.

Fixes #594.

PiperOrigin-RevId: 443813811