Use public API to test AWS KMS client in go.
Also, remove the check if the error message is errCredCSV, since errCredCSV is not in the public API and can't be used in tests anymore. Such checks are anyways discouraged, as the exact error message is not really part of the API and the user should not rely on it being the same value. And hence also our checks shouldn't do that.

PiperOrigin-RevId: 446734737