Adjust types of related to  singular jspb message getter functions.
Singular messages fields are typed to return `null` when the message is absent, this is inaccurate as they actually, mostly, return `undefined`​​​​​ when absent.  To clarify runtime semantics we will be adding `undefined` to the return type, so instead of returning `Foo|null` it will return `Foo|null|undefined`.  To keep code compatible we will insert type cast function calls at call sites that break with the change.  Occasionally we will adjust other types as well where inserting the cast is insufficient.

PiperOrigin-RevId: 452624526