Modify cross language tests to take the tests root path instead of Tink root.
Cross language tests now accept the path to the root of the cross language tests folder `TINK_CROSS_LANG_ROOT_PATH` instead of the root of the Tink directory `TINK_SRC_DIR`. Also, the search for a path no longer falls back to the next environment variable, but fails if a path is set and doesn't exist.

 - Replace TINK_SRC_DIR with TINK_CROSS_LANG_ROOT_PATH in testing/cross_language/util/testing_servers.py
 - Delete _path_util, replace _path_util.tink_root_path() with a more focused _root_folder() function local to testing/cross_language/util/testing_servers.py
 - Make _SERVER_PATHS and _JAVA_PATH relative to the root folder of cross lang tests rather than to the Tink's root directory.

PiperOrigin-RevId: 456320599