Rewrite PublicKeySign/Verify Factory tests.
These factories are deprecated and simply forward the call to KeysetHandle.getPrimitive. So having just one test case should be enough to check that the forward works.

The factories register the wrapper in case it has not been registered yet. I added a test for that. Since Registry.clear() is package private, I had to put that test case into its own test file.

PiperOrigin-RevId: 460213653