Rewrite DeterministicAeadFactoryTest.java.
DeterministicAeadFactory is deprecated and simply forwards the call to KeysetHandle.getPrimitive. So having just one test case should be enough to check that the forward works.

The factory registers the DeterministicAeadWrapper in case it has not been registered yet. I added a test for that. Since Registry.clear() is package private, I had to put that test case into its own test file.

PiperOrigin-RevId: 461603232