Add a CreateAead RPC to the Aead test service.
It should be possible to distinguish errors which occur because the Aead cannot be created, and errors which occur because encryption/decryption fails. (As in general, Tink aims for consistency that "if GetPrimitive succeeds, the result behaves the same in all languages)".

Tests then first need to call "CreateAead" to see if creation suceeeds.

PiperOrigin-RevId: 466956635