Move computation of region name into unit-tested code path.
Currently the code that calculates the region name is not run in unit tests, which makes the unit test accept invalid URIs. Using the real client, the code that calculates the region name is run and results in an IndexOutOfBoundsException.

This CL adds a test for this. The exception will be changed to IllegalArgumentException in the next cl.

Also, rename the local variable to keyId.

PiperOrigin-RevId: 469721127