Add deleteAt method to KeysetHandle.Builder and deprecate removeAt.
removeAt returns an KeysetHandle.Builder.Entry, which is usually not what the user wants. Instead, it is preferable to be able to chain the commands of the builder.

We add a new function and deprecate the old function to not break the API. Note that the name "deleteAt" is closer to the name of the method "delete" used by the KeysetManager.

Also, fix some lint warnings.

PiperOrigin-RevId: 472671792