ecdsa_sign_boringssl: Minor cleanup to DerToIeee
Passing a non-null first parameter to d2i_ECDSA_SIG isn't necessary
if you're using the return value anyway. (Which is the far more
straightforward version. The first parameter is quite confusing and
error-prone.)

Additionally, check for trailing data after calling d2i_ECDSA_SIG.
OpenSSL's d2i functions might only parse a prefix of the input,
advancing the input pointer past the bytes parsed. This is so you
can parse a series of structures concatenated together, but means
it's very easy to accidentally accept trailing data.

This is a no-op here because DerToIeee is only called on the output
of ECDSA_sign, which will always produce a valid signature. But in
case the function is ever copied elsewhere, add the extra check.

PiperOrigin-RevId: 477182844