Do not store the CC Config anymore in Objc TINKRegistryConfig objects.
The intention of these objects is:
1) The user creates the object.
2) The user sends the message "registerConfig" to the TINKConfig object:
"[TINKConfig registerConfig:createdConfig &error]"

The second message then registers everything.

However, in reality all the registering already happens when the "Config" object is created (e.g. in TINKMacConfig.mm -- see link below).

This means the whole rest is unnecessary and we can just delete it. We just register the configuration whenever it is asked.

PiperOrigin-RevId: 485262277