Use tinkerror.Fail() in hardcoded key template functions.
The exported functions in these packages that these changes apply to (e.g.
aead.AES128GCMKeyTemplate()) do not take any input and are expected not
to have any failure conditions. Adding explicit error handling for the
proto.Marshal() calls enforces this invariant so that if it doesn't hold,
it will fail during development (e.g. when running tests).

PiperOrigin-RevId: 485903201