Add toBigEndianBytesOfFixedLength to BigIntegerEncoding, and use it in SubtleUtil.
Also, remove the deprecation and inlining of bytes2Integer. I added that recently, but I think it's a mistake. It inlines a subtle function with an internal function, which is not what we want, since we don't want users of subtle functions to use internal functions. And since we currently don't have a good alternative for this function, and that function isn't really a problem, I think it's best to just leave it as it is.

If we decide to move BigIntegerEncoding into the public API, then we can deprecate and inline these function at that point. And migration for the user would be very easy.

PiperOrigin-RevId: 495038399