Delete aes_eax_aesni.
This is an implementation of AES EAX which does some verification against hardware errors. However, it is unused in Google, and has strictly speaking undefined behavior currently, because the allocator used in SecretData may allocate the _m128i objects unaligned.

This never had a BUILD.bazel or CMake, so it is extremely unlikely anyone uses it; I will not tag it as API change.

PiperOrigin-RevId: 501517535