Add tests for AndroidKeystoreKmsClient with fake AndroidKeystore Providers.
AndroidKeystoreKmsClient allows to inject a mock keystore objects, but that does not work as expected, since KeyGenerator is not mocked and the test still uses the real implementation. Also, hasKey resets the keystore object to the real object.

We therefore instead register fake AndroidKeystore Providers that gets used by AndroidKeystoreKmsClient.

This also resolves a TODO in AndroidKeystoreKmsClient.java.

PiperOrigin-RevId: 501652906