[C++] Monitoring for PRF Set
Wrap each PRF in the PRFSet with a `MonitoredPRF`. The `MonitoredPRF` will emit monitoring when enabled. The PRFSet now needs to also keep ownership of each MonitoredPRF. Because we already have a public facing API that returns a map<uint32,PRF*>, we can't remove that contract hence having a duplicate structure to keep track of the new MonitoredPRFs.

PiperOrigin-RevId: 502696442