Remove usage of SharedPrefKeysetReader in AndroidKeysetManager.
Instead of using a SharedPrefKeysetReader, we now first read keyset from the prefs file with a new function readKeysetFromPrefs(). It does this exactly as SharedPrefKeysetReader, but doesn't throw an error when the keyset does not exist and returns null instead.

readOrGenerateNewKeyset() is split into two functions, generateKeysetAndWriteToPrefs() and maybeDecryptAndParseKeyset().

To keep the exceptions the same, we use the BinaryKeysetReader and not TinkProtoKeysetFormat.

PiperOrigin-RevId: 504218201