Let hcvault reject key URI with bad format.
The hcvault client expects keys URIs to have a certain format. Unfortunately, if they don't have this format, calls to encrypt or decrypt may result in a panic. We now change that so that getAEAD will now return an error instead if the key URI is wrong.

Instead of creating the encryptionPath or decryptionPath when calling encrypt or decrypt, we pre-compute them and return an error if that doesn't work.

Also, we convert some methods that don't need access to the object's internals into pure functions.

PiperOrigin-RevId: 512533995