Add utility function to generate a random prefix for test files
When running tests with CMake, test files are created in /tmp. As a consequence, running tests in parallel is not possible due to the risk of tests using the same name for test files. This change adds a utility that generates unique test file name prefixes of the form `<test suite name>_<test name>_<random 32 bytes HEX>`.

PiperOrigin-RevId: 514344312