Let the Python implementation of KmsEnvelopeAead verify key_template.
This change has the effect that if an unknown key template is used, KmsEnvelopeAead will already fail when it is created, and not when it is used.

The C++ implementation also does this check, so it's better to also do it in Python, so that the behavior does not change when we migrate from wrapping the C++ version to the pure Python version.

At the same time, fix a lint warning.

PiperOrigin-RevId: 517961583