Implement native Python key managers KmsAeadKey and KmsEnvelopeAeadKey.
This is a refactoring, it should not change the behavior of the public API.

The KMS clients are now stored in a global list in Python. So we don't
use the C++-wrapped KMS clients registration anymore. Note that we do still use the C++ KMS clients, and wrap them into Python clients.

The next step will be to implement the clients in Python.

PiperOrigin-RevId: 518294262