Implement an option to specify the name used in the AWS KMS EncryptionContext.
This provides a solution to the ciphertext incompatibility problem described in
#568. In particular, the EncryptionContextName can now be set to
"associatedData", which aligns with the Tink KMS extensions in other languages.
It is made an option to avoid stranding old ciphertexts.

To facilitate this new option, a new awsClient factory function,
NewClientWithOptions(), is implemented which accepts functional options.
Options are introduced to cover all of the capabilities of the existing factory
functions, which are now deprecated.

When using NewClientWithOptions() the default EncryptionContextName is
AssociatedData. When using any of the deprecated factory functions, the default
EncryptionContextName is LegacyAdditionalData. This provides balance between not
breaking existing usages and encourages migration from the deprecated functions.

PiperOrigin-RevId: 537147115