Remove parseKeyData from PrivilegedRegistry.
This is unused. The class PrivilegedRegistry has the comment "This is internal and should not be used externally.", so it is clearly marked that it is not part of the public API. And I think it is very unlikely that anybody uses this by accident. So I think we can just remove it.

#tinkApiChange

PiperOrigin-RevId: 543701753