Change Bytes.copyFrom such that it is not illegal to call it with bad ranges.
When we create a Bytes object from e.g. a ciphertext (which is convenient to e.g. check the prefix) it should be possible to write "Bytes.copyFrom(ciphertext, 0, 5)" even if the ciphertext is too short. In this case, we just return the remainder -- when we later call "Equals" it will do the right thing.

PiperOrigin-RevId: 557516980