Remove the ability of the registry to store a name -> parameters mapping.
Everything is now stored in the MutableParametersRegistry and should be directly used at registration.

Note that this requires KeyTypeManagers to be internal for it to be safe. Luckily, they are.

This will change behavior if users call "register" with "newKeyAllowed = false" (in which case it now also registers the named parameters) and if users create a keymanager themselves (which should be impossible since the constructors are private).

PiperOrigin-RevId: 565269901